@@ -2,7 +2,6 @@
   "name": "pocket-monorepo",
   "private": true,
   "scripts": {
-    "prebuild": "dotenv -- turbo run prebuild",
     "build": "dotenv -- turbo run build",
     "clean": "rm -rf .turbo node_modules packages/**/.turbo packages/**/dist packages/**/node_modules servers/**/logs servers/**/.turbo servers/**/dist servers/**/node_modules lambdas/**/logs lambdas/**/.turbo lambdas/**/dist lambdas/**/node_modules infrastructure/**/dist infrastructure/**/node_modules infrastructure/**/.turbo infrastructure/**/cdktf.out",
     "dev": "dotenv -- turbo run dev --concurrency 50",
@@ -15,7 +14,6 @@
     "prepare": "husky",
     "semantic-release": "dotenv -- turbo run semantic-release",
     "synth": "dotenv -- turbo run synth",
-    "pretest-integrations": "dotenv -- turbo run pretest-integrations",
     "test": "dotenv -- turbo run test",
     "test-integrations": "dotenv -- turbo run test-integrations"
   },
@@ -26,18 +24,18 @@
   },
   "dependencies": {
     "dotenv-cli": "latest",
-    "husky": "9.1.1"
+    "husky": "9.1.4"
   },
   "devDependencies": {
-    "@commitlint/cli": "19.3.0",
+    "@commitlint/cli": "19.4.0",
     "@commitlint/config-conventional": "^19.2.2",
     "@pocket-tools/eslint-config": "workspace:*",
-    "syncpack": "^12.3.3",
+    "syncpack": "^12.4.0",
     "tsconfig": "workspace:*",
-    "turbo": "^1.13.4"
+    "turbo": "^2.1.0"
   },
-  "packageManager": "pnpm@9.1.4",
+  "packageManager": "pnpm@9.9.0",
   "engines": {
-    "node": "^20.15"
+    "node": "^20.16"
   }
 }
