@@ -1,3 +1,7 @@
 import packages from '@pocket-tools/eslint-config/packages';
 import tseslint from 'typescript-eslint';
-export default tseslint.config(...packages);
+export default tseslint.config(...packages, {
+  rules: {
+    '@typescript-eslint/no-require-imports': ['error', { allow: ['/logger$'] }],
+  },
+});
