@@ -1,7 +1,6 @@
 import { Strategy } from 'unleash-client';
 import { UnleashContext, RecItUserProfile } from '../../graphql/typeDefs';
 import { normalizedStrategyValue } from 'unleash-client/lib/strategy/util';
-import { SessionIdError } from '../../utils/customErrors';
 import * as Sentry from '@sentry/node';
 
 export class HasUserModel extends Strategy {
@@ -13,12 +12,7 @@ export class HasUserModel extends Strategy {
     try {
       const groupId = parameters.groupId || context.featureToggle || '';
       const percentage = Number(parameters.rollout);
-      const stickinessId = context.userId || context.sessionId;
-
-      // If there is no sticky id
-      if (!stickinessId || !stickinessId?.length) {
-        throw new SessionIdError('No Stickiness ID Provided');
-      }
+      const stickinessId = context.userId || context.sessionId || '';
 
       // Users without a user profile are not eligible.
       const userProfileString = context?.properties?.recItUserProfile;
