@@ -1,7 +1,6 @@
 import { Strategy } from 'unleash-client';
 import { UnleashContext } from '../../graphql/typeDefs';
 import { normalizedStrategyValue } from 'unleash-client/lib/strategy/util';
-import { SessionIdError } from '../../utils/customErrors';
 import * as Sentry from '@sentry/node';
 
 export class UserInLocaleStrategy extends Strategy {
@@ -13,12 +12,7 @@ export class UserInLocaleStrategy extends Strategy {
     try {
       const groupId = parameters.groupId || context.featureToggle || '';
       const percentage = Number(parameters.rollout);
-      const stickinessId = context.userId || context.sessionId;
-
-      // If there is no sticky id
-      if (!stickinessId || !stickinessId?.length) {
-        throw new SessionIdError('No Stickiness ID Provided');
-      }
+      const stickinessId = context.userId || context.sessionId || '';
 
       // Check for locale
       const locale = context?.properties?.locale;
