@@ -4,24 +4,13 @@ import { initSentry } from '@pocket-tools/sentry';
 initSentry({
   ...config.sentry,
   debug: config.sentry.environment == 'development',
-  skipOpenTelemetrySetup: true,
 });
-//this must run before all imports and server start but after sentry
-//so open-telemetry can patch all libraries that we use
-import { nodeSDKBuilder } from '@pocket-tools/tracing';
+
+import { startServer } from './apollo';
 import { serverLogger } from '@pocket-tools/ts-logger';
 
-nodeSDKBuilder({
-  host: config.tracing.host,
-  serviceName: config.tracing.serviceName,
-  release: config.sentry.release,
-  logger: serverLogger,
-  addSentry: true,
-}).then(async () => {
-  await startServer(config.app.serverPort);
+startServer(config.app.serverPort).then(() => {
   serverLogger.info(
     `🚀 Public server ready at http://localhost:${config.app.serverPort}`,
   );
 });
-
-import { startServer } from './apollo';
