@@ -7,22 +7,12 @@ import { initSentry } from '@pocket-tools/sentry';
 initSentry({
   ...config.sentry,
   debug: config.sentry.environment == 'development',
-  skipOpenTelemetrySetup: true,
 });
-import { nodeSDKBuilder } from '@pocket-tools/tracing';
 import { serverLogger } from '@pocket-tools/ts-logger';
+import { startServer } from './apollo/server';
 
-nodeSDKBuilder({
-  host: config.tracing.host,
-  serviceName: config.tracing.serviceName,
-  release: config.sentry.release,
-  logger: serverLogger,
-  addSentry: true,
-}).then(async () => {
-  const { url } = await startServer(config.app.serverPort);
+startServer(config.app.serverPort).then(({ url }) => {
   serverLogger.info(
     `🚀 Public server ready at http://localhost:${config.app.serverPort}${url}`,
   );
 });
-
-import { startServer } from './apollo/server';
