@@ -6,7 +6,7 @@
  * - CircleCI config - Create a job to the function test npm script from above
  */
 
-import nock, { cleanAll } from 'nock';
+import nock, { cleanAll, restore } from 'nock';
 import { getRedis } from '../../cache';
 import { VideoType, Videoness } from '../../__generated__/resolvers-types';
 import { startServer } from '../../apollo/server';
@@ -104,6 +104,7 @@ describe('Marticle integration ', () => {
     await server.stop();
     await getRedis().disconnect();
     cleanAll();
+    restore();
   });
 
   it('should return marticle text for the given url', async () => {
@@ -148,7 +149,7 @@ describe('Marticle integration ', () => {
       },
     ];
 
-    expect(res.body.data).not.toBeUndefined;
+    expect(res.body.data).not.toBeUndefined();
     const marticle = res.body.data.itemByUrl.marticle;
     expect(marticle.length).toBeGreaterThan(0);
     expect(marticle).toStrictEqual(expected);
@@ -196,7 +197,7 @@ describe('Marticle integration ', () => {
     const res = await request(app)
       .post(graphQLUrl)
       .send({ query: print(GET_ITEMS_BY_URL), variables });
-    expect(res.body.data).not.toBeUndefined;
+    expect(res.body.data).not.toBeUndefined();
     const marticle = res.body.data.itemByUrl.marticle;
     expect(marticle.length).toBeGreaterThan(0);
     expect(marticle).toStrictEqual([
@@ -263,7 +264,7 @@ describe('Marticle integration ', () => {
     const res = await request(app)
       .post(graphQLUrl)
       .send({ query: print(GET_ITEMS_BY_URL), variables });
-    expect(res.body.data).not.toBeUndefined;
+    expect(res.body.data).not.toBeUndefined();
     const marticle = res.body.data.itemByUrl.marticle;
     expect(marticle.length).toBeGreaterThan(0);
     expect(marticle).toStrictEqual([
@@ -319,7 +320,7 @@ describe('Marticle integration ', () => {
       .post(graphQLUrl)
       .send({ query: print(GET_ITEMS_BY_URL), variables });
 
-    expect(res.body.data).not.toBeUndefined;
+    expect(res.body.data).not.toBeUndefined();
     const marticle = res.body.data.itemByUrl.marticle;
     expect(marticle).toStrictEqual([]);
   });
@@ -358,7 +359,7 @@ describe('Marticle integration ', () => {
     const res = await request(app)
       .post(graphQLUrl)
       .send({ query: print(GET_ITEMS_BY_URL), variables });
-    expect(res.body.data).not.toBeUndefined;
+    expect(res.body.data).not.toBeUndefined();
     const marticle = res.body.data.itemByUrl.marticle;
     expect(marticle.length).toBeGreaterThan(0);
     expect(marticle).toStrictEqual([
