@@ -1,4 +1,4 @@
-import { cleanAll } from 'nock';
+import { cleanAll, restore } from 'nock';
 import { getRedis } from '../../cache';
 import { startServer } from '../../apollo/server';
 import { ApolloServer } from '@apollo/server';
@@ -111,6 +111,7 @@ describe('readerSlug', () => {
     await server.stop();
     await getRedis().disconnect();
     cleanAll();
+    restore();
     await readitlaDB.destroy();
     await sharesInit().destroy();
     jest.restoreAllMocks();
