@@ -3,25 +3,13 @@ import { initSentry } from '@pocket-tools/sentry';
 initSentry({
   ...config.sentry,
   debug: config.sentry.environment == 'development',
-  skipOpenTelemetrySetup: true,
 });
 
-//this must run before all imports and server start but before sentry
-//so open-telemetry can patch all libraries that we use
-import { nodeSDKBuilder } from '@pocket-tools/tracing';
 import { serverLogger } from '@pocket-tools/ts-logger';
+import { startServer } from './server';
 
-nodeSDKBuilder({
-  host: config.tracing.host,
-  serviceName: config.tracing.serviceName,
-  release: config.sentry.release,
-  logger: serverLogger,
-  addSentry: true,
-}).then(async () => {
-  await startServer(config.app.port);
+startServer(config.app.port).then(() => {
   serverLogger.info(
     `🚀 Public server ready at http://localhost:${config.app.port}`,
   );
 });
-
-import { startServer } from './server';
